Rejig files and directory structure
